+++ laboratory/gradle-plugin/src/main/kotlin/io/mehow/laboratory/gradle/FeatureFlagInput.kt
@@ -1,23 +1,19 @@
 package io.mehow.laboratory.gradle
 
 import com.squareup.kotlinpoet.ClassName
-import io.mehow.laboratory.generator.Deprecation
 import io.mehow.laboratory.generator.FeatureFlagModel
-import io.mehow.laboratory.generator.FeatureFlagOption
-import io.mehow.laboratory.generator.Supervisor
-import io.mehow.laboratory.generator.Visibility.Internal
-import io.mehow.laboratory.generator.Visibility.Public
-import io.mehow.laboratory.gradle.DeprecationLevel.Warning
+import io.mehow.laboratory.generator.Visibility
 import org.gradle.api.Action
+import java.io.Serializable
 
 /**
  * Representation of a generated feature flag. It must have at least one value and exactly one default value.
  */
 public class FeatureFlagInput internal constructor(
   private val name: String,
-  private val packageNameProvider: () -> String,
-  private val supervisor: (() -> Supervisor)? = null,
-) {
+  packageNameProvider: PackageNameProvider,
+  private val supervisor: SupervisorInput?,
+) : Serializable {
   /**
    * Sets whether the generated feature flag should be public or internal.
    */
@@ -26,7 +22,11 @@ public class FeatureFlagInput internal constructor(
   /**
    * Sets package name of the generated feature flag. Overwrites any previously set values.
    */
-  public var packageName: String? = null
+  public var packageName: String?
+    get() = packageNameProvider.value
+    set(value) = packageNameProvider.setValue(value)
+
+  private val packageNameProvider = PackageNameProvider(packageNameProvider)
 
   /**
    * Sets description of the generated feature flag.
@@ -38,12 +38,14 @@ public class FeatureFlagInput internal constructor(
    */
   public var key: String? = null
 
-  private val options: MutableList<FeatureFlagOption> = mutableListOf()
+  private val options = mutableListOf<FeatureFlagOptionInput>()
+
+  private val childFeatureInputs = mutableListOf<ChildFeatureFlagsInput>()
 
   /**
    * Adds a feature option.
    */
-  public fun withOption(name: String): Unit = withOption(name) { }
+  public fun withOption(name: String): Unit = withOption(name, action = {})
 
   /**
    * Adds a feature option and configures features flags supervised by it.
@@ -51,14 +53,13 @@ public class FeatureFlagInput internal constructor(
   public fun withOption(
     name: String,
     action: Action<ChildFeatureFlagsInput>,
-  ): Unit =
-    withOption(name, isDefault = false, action)
+  ): Unit = withOption(name, isDefault = false, action = action)
 
   /**
    * Adds a feature value that will be used as a default value.
    * Exactly one value must be set with this method.
    */
-  public fun withDefaultOption(name: String): Unit = withDefaultOption(name) { }
+  public fun withDefaultOption(name: String): Unit = withDefaultOption(name, action = {})
 
   /**
    * Adds a feature value that will be used as a default value and configures features flags supervised by it.
@@ -67,68 +68,68 @@ public class FeatureFlagInput internal constructor(
   public fun withDefaultOption(
     name: String,
     action: Action<ChildFeatureFlagsInput>,
-  ): Unit =
-    withOption(name, isDefault = true, action)
-
-  private val childFeatureInputs = mutableListOf<ChildFeatureFlagsInput>()
+  ): Unit = withOption(name, isDefault = true, action = action)
 
   private fun withOption(
     name: String,
     isDefault: Boolean,
     action: Action<ChildFeatureFlagsInput>,
   ) {
-    val option = FeatureFlagOption(name, isDefault)
+    val option = FeatureFlagOptionInput(name, isDefault)
+    val supervisor = SupervisorInput(this, option)
+    childFeatureInputs += ChildFeatureFlagsInput(packageNameProvider, supervisor).apply(action::execute)
     options += option
-    val packageNameProvider = { packageName ?: packageNameProvider() }
-    val supervisorBuilder = { Supervisor(toModel(), option) }
-    childFeatureInputs += ChildFeatureFlagsInput(packageNameProvider, supervisorBuilder).let { input ->
-      action.execute(input)
-      return@let input
-    }
   }
 
-  private val sources: MutableList<FeatureFlagOption> = mutableListOf()
+  private val sources = mutableListOf<FeatureFlagOptionInput>()
 
   /**
    * Adds a feature flag source. Any sources that are named "Local", or any variation of this word,
    * will be filtered out.
    */
-  public fun withSource(name: String) {
-    sources += FeatureFlagOption(name)
-  }
+  public fun withSource(name: String): Unit = withSource(name, isDefault = false)
 
   /**
    * Adds a feature flag source that will be used a default source. Any sources that are named "Local",
    * or any variation of this word, will be filtered out.
    * At most one value can be set with this method.
    */
-  public fun withDefaultSource(name: String) {
-    sources += FeatureFlagOption(name, isDefault = true)
+  public fun withDefaultSource(name: String): Unit = withSource(name, isDefault = true)
+
+  private fun withSource(
+    name: String,
+    isDefault: Boolean,
+  ) {
+    val option = FeatureFlagOptionInput(name, isDefault)
+    sources += option
   }
 
-  private var deprecation: Deprecation? = null
+  private var deprecation: DeprecationInput? = null
 
   /**
    * Annotates a feature flag as deprecated.
    */
   @JvmOverloads public fun deprecated(
     message: String,
-    level: DeprecationLevel = Warning,
+    level: DeprecationLevel = DeprecationLevel.Warning,
   ) {
-    deprecation = Deprecation(message, level.kotlinLevel)
+    deprecation = DeprecationInput(message, level.kotlinLevel)
   }
 
-  private fun toModel() = FeatureFlagModel(
-    visibility = if (isPublic) Public else Internal,
-    className = ClassName(packageName ?: packageNameProvider(), name),
-    options = options,
-    sourceOptions = sources,
+  internal fun toModel() = FeatureFlagModel(
+    visibility = if (isPublic) Visibility.Public else Visibility.Internal,
+    className = ClassName(packageNameProvider.value.orEmpty(), name),
+    options = options.map(FeatureFlagOptionInput::toModel),
+    sourceOptions = sources.map(FeatureFlagOptionInput::toModel),
     key = key,
     description = description.orEmpty(),
-    deprecation = deprecation,
-    supervisor = supervisor?.invoke(),
+    deprecation = deprecation?.toModel(),
+    supervisor = supervisor?.toModel(),
   )
 
-  internal fun toModels(): List<FeatureFlagModel> =
-    listOf(toModel()) + childFeatureInputs.flatMap(ChildFeatureFlagsInput::toModels)
+  internal fun toModels() = listOf(toModel()) + childFeatureInputs.flatMap(ChildFeatureFlagsInput::toModels)
+
+  internal companion object {
+    private const val serialVersionUID = 0L
+  }
 }