+++ laboratory/gradle-plugin/src/main/kotlin/io/mehow/laboratory/gradle/OptionFactoryInput.kt
@@ -5,13 +5,14 @@ import io.mehow.laboratory.generator.FeatureFlagModel
 import io.mehow.laboratory.generator.OptionFactoryModel
 import io.mehow.laboratory.generator.Visibility.Internal
 import io.mehow.laboratory.generator.Visibility.Public
+import java.io.Serializable
 
 /**
  * Representation of a generated option factory that is aware of feature flags.
  */
 public class OptionFactoryInput internal constructor(
-  private val packageNameProvider: () -> String,
-) {
+  packageNameProvider: PackageNameProvider,
+) : Serializable {
   /**
    * Sets whether the generated option factory should be public or internal.
    */
@@ -20,11 +21,19 @@ public class OptionFactoryInput internal constructor(
   /**
    * Sets package name of the generated option factory. Overwrites any previously set values.
    */
-  public var packageName: String? = null
+  public var packageName: String?
+    get() = packageNameProvider.value
+    set(value) = packageNameProvider.setValue(value)
+
+  private val packageNameProvider = PackageNameProvider(packageNameProvider)
 
   internal fun toModel(features: List<FeatureFlagModel>) = OptionFactoryModel(
     visibility = if (isPublic) Public else Internal,
-    className = ClassName(packageName ?: packageNameProvider(), "GeneratedOptionFactory"),
+    className = ClassName(packageNameProvider.value.orEmpty(), "GeneratedOptionFactory"),
     features = features,
   )
+
+  internal companion object {
+    private const val serialVersionUID = 0L
+  }
 }